Clustering II
These changes will allow Jentrata to run as a cluster. The cluster can
be build up out of several Jentrata instances on top of one Jentrata
database. In this version only Postgresql is supported, but changes in
the appropiate property files will allow to run it on top of other
DBMSes.

After installing this implementation make sure it will look at the
central postgresql database. This can be set in the property file:
./plugins/hk.hku.cecid.ebms/conf/hk/hku/cecid/ebms/spa/conf/ebms.module.
xml under the home directory of Jentrata. You need to change the line:
<parameter name="url" value="jdbc:postgresql://localhost:5432/ebms" />
by replacing 'localhost' into the address of the central postgresql
database.

On other important setting is to make sure the 'hostname' can be
translated into an ip address that's unique for the Jentrata server.