Cluster II Audit
Additional code for the Cluster II. This code will audit the central
database for hosts which are no longer available in the cluster but
left messages unprocessed.

This audit is part of each host within the cluster. The audit code
itself is aware of the other hosts within the cluster. This is done via
a newly created table called 'cluster'.

In the cluster table each host will be represented by its ip-address,
status (active, inactive, auditor) and a timestamp telling each host
when it is allowed to run their audit.