Update ClusterAudit.java
The ClusterAudit is causing errors in the corvus.log file, each time it is checking if a host is available or not. This is because it is sending an empty http request to the listening service of the host which is checked. This empty request is seen as an incorrect ebMS request and therefor causes an error in the corvus.log file. This change will update the hostIsAvailable() method to perform a portscan in stead of a http request.