Remove DuplicateElimination from Acknowledgement
This is the change needed for issue #62 Remove DuplicateElimination from Acknowledgement. The original code result in compatibility problems with a number of ebMS v2.0 implementations. Since the addition of the DuplicateElimination setting as part of the acknowledgement is not allowed as stated in the ebMS v2.0 standard provided by OASIS this should be removed from Jentrata.