Dockfile does not build
In the Dockfile there are some errors which prevents it from building a Jentrata container. The base, FROM, container does not exist. It needs to be changed into jentrata/jentrata-msh. In this container the JAVA_VERSION is set to 10, but version 10 is no longer available in the debian repository. By removing the openjdk install line it will not try to install it.